Move interface to documented class
